add new fields to package, update tests
- added stars and hasTestDirectory to packages
- updated tests to fix broken tests
- moves tests -> test, as this is more standard
- updated info tests, working towards #48
