+++ package.json
@@ -18,7 +18,7 @@
     "debug": "npm run nargs ncq/main -- --log",
     "analytics": "npm run nargs ncq/main -- --usage",
     "searchless": "npm run nargs ncq/main -- --usage --searchless",
-    "test": "mocha tests --max-old-space-size=4096",
+    "test": "mocha test --max-old-space-size=4096",
     "coverage": "nyc npm run test",
     "info": "mocha tests/info --max-old-space-size=4096"
   },
