+++ test/ncq-test.js
@@ -1,6 +1,5 @@
 require("mocha");
 var assert = require("assert");
-const path = require("path");
 const NCQ = require("../ncq/core/ncq");
 
 /**
@@ -9,24 +8,19 @@ const NCQ = require("../ncq/core/ncq");
 describe("NCQ", function () {
     var ncqService;
     describe("functions", function(){
-        it("should initialize object", function(){
-            ncqService = new NCQ();
-            ncqService.initialize();
+        it("should initialize object", async function(){
+            ncqService = new NCQ({recordLimit: 1000});
+            await ncqService.initialize();
 
             assert(ncqService !== undefined);
             assert(ncqService.state !== undefined);
         }).timeout(0); //no timeout for loading
-
-        it("should search for snippets by task", function(){
-            var snippets = ncqService.snippetsByTask("read file");
-            assert(snippets.length > 0);
-        });
         it("should search for snippets by package", function(){
-            var snippets = ncqService.snippetsByPackages(["acute"]);
+            var snippets = ncqService.snippetsByPackages(["111-hello-world"]);
             assert(snippets.length > 0);
         });
         it("should search for packages by task", function(){
-            var packages = ncqService.packagesByTask("read file");
+            var packages = ncqService.packagesByTask("utility");
             assert(packages.length > 0);
         });
     });
